pyproject.toml
[tool.poetry]
name = "qa-guru-python-10-20"
version = "0.1.0"
description = ""
authors = ["Dimitri Maksimov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "2.31.0"
jsonschema = "4.21.1"
pytest = "8.1.1"
python-dotenv = "1.0.1"
selene = "2.0.0rc9"
allure-pytest = "2.13.4"
curlify = "^2.2.1"

[tool.pytest.ini_options]
addopts = [
    "--alluredir", "allure-results",
    "--clean-alluredir",
    "-s",
    "-v",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
